Resets the remembered previously selected value when setting new value
When we set a new value for a single select we internally clear the old value and then set the new value. The fields that stored the previous value (for restoring it when the user closes the suggestions without choosing a new value) were not cleared in that case which made race conditions in the event handling and their order more likely.

Fixes: SIRI-490