Tries to keep suggestions more consistent
This change clears the current suggestions more often to prevent weird behaviour with hidden suggestions.
The alternative would have been to check whether the suggestions are currently visible everywhere where we access them, which could bring unwanted side effects when forgotten somewhere in the future.
It is highly unlikely that we reuse the suggestions present when hidden anyways.

The issue that inspired this change was the following:
When adding a custom token (on multi select fields supporting it) it was possible to create a duplicate token simply by hitting the Enter key a second time (As the Enter key would automatically select the single hidden custom token that was not removed)

Fixes: OX-8762