Cancels focusout handling on active instead of highlighted suggestion
This was previously introduced in the following PR:
https://github.com/scireum/token-autocomplete/pull/52

The reason for this was a race condition (or rather unfortunate event timing) in regards to the click on suggestions and the focus out of the input field.

But checking for a highlighted (hover or keyboard) suggestion was too broad. It resulted in the suggestions staying open when a suggestion is highlighted and Tab key is used to jump to the next field (when the field has option `enableTabulator` set to `false`).

We now solve both issues by only preventing hiding the suggestions on focusout of the input field when a suggestion is active (clicked) simultaneously.

Fixes: OX-8732