Allows to prevent adding selected suggestion as token
By setting the data-attribute `becomesToken` to false a user of the library can prevent the selected suggestion from turning into a token on the field.
By also registering a `suggestion-selected` event listener on the suggestion he can then add custom handling for the selection of the token (regardless whether the suggestion selected via keyboard or mouse).

Fixes: OX-10302