Hide the suggestion when the arrow button is pressed on an active field
For this we had to turn the pseudo element into a button as no event listeners can be registered on pseudo elements.
We also have to mark the button in some way when the input loses focus because of the DOM event order. Focusout is triggered before click so the input would have already lost focus when the click handler runs and it would set the focus again.

Fixes: SIRI-981