Uses fixed positioning to prevent overflow in parent DIV while showing suggestions
This solves (after a lot of fiddling around) a problem where the suggestions list could cause the container to overflow (and show scroll bars) when it does not have the capacity for all suggestions.
Which caused weird behaviour and look&feel when e.g. using a single token autocomplete field in a modal.

To combat this we position the suggestions container fixed, decoupling it from normal page flow and calculate the correct position relative to the input field.
This makes it necessary to close the suggestion when the user scrolls or resizes the window, as otherwise the suggestions and input field would separate visually.

Fixes: SIRI-1128