Add capability to resolve label for a given value during initial setup
Sometimes the display label of an initial token / selection may not be easy to retrieve for field setup. In such cases we now allow to specify an URI that is called for each initial token during setup. This is entirely optional.

To keep the code somewhat readable we now depend on the Promise feature and require a polyfill when used in IE11.

Fixes: SIRI-1129